Decouple player setting from Multiplayer.class Player direction now sent over network

+++ src/main/java/project_16x16/components/AnimationComponent.java
@@ -92,8 +92,8 @@ public class AnimationComponent {
 	*
 	* @return the current frame as a float
 	**/
-	public float getFrameID() {
-		return currentFrame;
+	public int getFrameID() {
+		return (int) currentFrame;
 	}
 	
 	/**
